# s2: Do 2 overlapping match checks