… s2/encodeblock_amd64.s