Mainly faster on long matches, but has little to no regression on short matches.

```
λ s2c -cpu=1 --bench=5 \compress\github-june-2days-2019.json

BEFORE: Compressing... 6273951764 -> 944876986 [15.06%]; 9.43s, 634.5MB/s
AFTER:  Compressing... 6273951764 -> 944876986 [15.06%]; 9.269s, 645.5MB/s

λ s2c -cpu=1 --bench=5 \compress\nyc-taxi-data-10M.csv
BEFORE: Compressing... 3325605752 -> 884713054 [26.60%]; 8.296s, 382.3MB/s
AFTER:  Compressing... 3325605752 -> 884713054 [26.60%]; 8.173s, 388.1MB/s

λ s2c -cpu=1 --bench=5 \compress\enwik9
BEFORE: Compressing... 1000000000 -> 416582393 [41.66%]; 3.706s, 257.3MB/s
AFTER:  Compressing... 1000000000 -> 416582393 [41.66%]; 3.636s, 262.3MB/s
```